minor adjustments

+++ src/gifs.py
@@ -3,7 +3,7 @@ import glob
 import os
 import sys
 
-def gifs_from_png(list_of_filenames, save_path, duration=0.2):
+def gifs_from_png(list_of_filenames, save_path, duration=0.2, remove=True):
     images = []
 
     for filename in list_of_filenames:
@@ -12,6 +12,7 @@ def gifs_from_png(list_of_filenames, save_path, duration=0.2):
 
     imageio.mimsave(save_path, images, duration=duration)
 
-    for f in list_of_filenames:
-        if os.path.exists(f):
-            os.remove(f)
+    if remove:
+        for f in list_of_filenames:
+            if os.path.exists(f):
+                os.remove(f)
